Use Buildkite to select the version for release
The Buildkite pipeline will ask what version we want to release.
On the prompt only the **5 latest versions** are displayed. When we need
to build a new version the pipeline step also needs to be updated to reflect
the change.

Use [Docker Compose Buildkite plugin](https://github.com/buildkite-plugins/docker-compose-buildkite-plugin) to build images

Because we have `docker-compose` v1 installed, for the `docker-compose`
plugin we have to specify the `cli_version=1` argument to use the
`docker-compose` command, instead of the newer `docker compose`.

This also updates the pipeline workflow:

1. `.buildkite/pipeline.yml` is now the entry point for new builds.
2. `.buildkite/pipeline.master.yml` used for the branch `master`, can
   build and release images and create a Github release.
3. `./buildkite/pipeline.pull_request.yml` used for Pull Requests, can
   build and release images, doesn't create a Github release.